legacy/requirements.txt
# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3

# Geospatial
rasterio>=1.3
geopandas>=0.13
shapely>=2.0
pyproj>=3.5
rioxarray>=0.14
xarray>=2023.1

# Modeling
xgboost>=2.0
shap>=0.43

# Visualization
matplotlib>=3.7
seaborn>=0.12
contextily>=1.3

# Configuration & utilities
pyyaml>=6.0
tqdm>=4.65
requests>=2.31

# Optional: parallel/large-grid handling
joblib>=1.3
dask>=2023.10

# Optional: Google Earth Engine for MODIS (recommended for fast access)
# earthengine-api>=0.1.380
